feat: extract money managed chart to separate component
