+++ lang/ca.json
@@ -712,7 +712,6 @@
   "Crypto": "Crypto",
   "Currency": "Currency",
   "currency": "divisa",
-  "CW9iCu": "See historical",
   "Date.BeforeShort": "Before {date, date, short}",
   "Date.DateRange": "{dateFrom, date, short} to {dateTo, date, short}",
   "Date.SinceShort": "Since {date, date, short}",
