+++ lang/zh.json
@@ -712,7 +712,6 @@
   "Crypto": "加密货币",
   "Currency": "货币",
   "currency": "货币",
-  "CW9iCu": "See historical",
   "Date.BeforeShort": "在 {date, date, short} 之前",
   "Date.DateRange": "{dateFrom, date, short} 至 {dateTo, date, short}",
   "Date.SinceShort": "自 {date, date, short}",
